Swap menu-new to menu
This completes the full refactor of converting the sidebar menu
into a function component
